backend/requirements.txt
# StoryRag v2.0 后端依赖

# FastAPI框架
fastapi>=0.110.0
uvicorn>=0.29.0

# 原有项目依赖
# PyTorch CUDA版本 (GPU加速)
torch>=2.7.0
torchaudio>=2.7.0
torchvision>=0.22.0

# LangChain框架
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# 向量嵌入和检索
sentence-transformers>=2.7.0
transformers>=4.39.0
accelerate>=0.28.0
safetensors>=0.4.0

# 文档处理
PyPDF2>=3.0.0
pypdf>=4.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0

# 数据处理和科学计算
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# 环境变量管理
python-dotenv>=1.0.0

# 日志和配置
loguru>=0.7.0
pydantic>=2.5.0

# HTTP请求
httpx>=0.27.0
requests>=2.31.0

# JSON和数据序列化
orjson>=3.9.0

# 向量数据库工具
faiss-cpu>=1.7.0
pinecone-client>=2.2.0
